third_party/blink/web_tests/external/wpt/html/canvas/offscreen/text/2d.text.drawing.style.fontVariant.settings.worker-expected.txt

This is a testharness.js-based test.
FAIL Testing basic functionalities of fontVariant for canvas assert_equals: ctx.fontVariantCaps === "titling-caps" (got normal[string], expected titling-caps[string]) expected "titling-caps" but got "normal"
Harness: the test ran to completion.
